Create script
Created the first version of the script.

Supports the registration of any number of railroad crossings, each with a defined closing and opening sequence ("routine") consisting of any number of commands. Registered crossings are closed whenever they are open and a train approaches, and only opened once no train is approaching anymore. The number of approaching trains can be saved in EEP save slots of the user's choice in order for that data to survive re-loads of the script or the whole map file.