Add new crossing definition procedure
Added a new, function-chaining-based way of defining railroad crossings. The old way via the setup function remains available, but has been marked as deprecated and will be removed in the next major release.

Moved the update_and_get_trains and process_queue functions from the SW object to the UTILS object, so as to not expose them to the user.

Version number bumped by a minor release.