Support reversing sequence
Introduced the possibility to define a "reversing" sequence for crossings. If a crossing is activated anew while another sequence - most likely the opening sequence - is still ongoing, that sequence is cancelled, and the "reversing" sequence is immediately executed instead. This allows for the simulation of the special "another train coming" warnings on UK crossings, or the immediate re-lowering of the barriers on WSSB crossings.